TASK: replaced old routes
Replaced old routes scripts with a more modular approach
